brkn: test  v3.0.1 -> gLit_GUI

+++ 3dLand/GlLand/gLit_GUI/src/headers/font.h
@@ -17,7 +17,7 @@ struct Char {
   uint ID;            // TextureID
   ivec2 Size;         // GlyphSize
   ivec2 Bearing;      // Offset from baseline from L to R 
-  long Advance;       // Offset to advance to next glyph
+  int Advance;       // Offset to advance to next glyph
 };
 
 class Font {
